move compare save dialog component
