+++ packages/web-app-user-management/src/components/Users/SideBar/EditPanel.vue
@@ -48,7 +48,7 @@
 <script>
 import * as EmailValidator from 'email-validator'
 import UserInfoBox from './UserInfoBox.vue'
-import CompareSaveDialog from '../../CompareSaveDialog.vue'
+import CompareSaveDialog from 'web-pkg/src/components/sidebar/CompareSaveDialog.vue'
 
 export default {
   name: 'EditPanel',
