+++ packages/web-app-user-management/src/components/Users/SideBar/GroupAssignmentsPanel.vue
@@ -47,7 +47,7 @@
 </template>
 <script>
 import UserInfoBox from './UserInfoBox.vue'
-import CompareSaveDialog from '../../CompareSaveDialog.vue'
+import CompareSaveDialog from 'web-pkg/src/components/sidebar/CompareSaveDialog.vue'
 
 export default {
   name: 'GroupAssignmentsPanel',
